cloud/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
paho-mqtt>=1.6.1
python-socketio>=5.9.0
